requirements.txt
# Core JAX/Flax dependencies
jax[cuda12]>=0.4.20
flax>=0.8.0
optax>=0.1.9

# TensorFlow for data loading
tensorflow>=2.15.0
tensorflow-datasets>=4.9.0

# Core Python libraries
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and logging
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Machine learning utilities
scikit-learn>=1.3.0

# Image processing
Pillow>=10.0.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
